server-python/requirements-prod.txt
# ===========================================
# OpenCraft Backend - 生产环境依赖
# ===========================================
# 仅包含运行时必需的依赖，最小化镜像体积

# Web 框架
fastapi==0.110.0

# ASGI 服务器（不含可选扩展）
uvicorn==0.27.1

# 数据库 ORM + 异步支持
sqlalchemy[asyncio]==2.0.25

# SQLite 异步驱动
aiosqlite==0.19.0

# 数据验证
pydantic==2.6.1

# LLM API 客户端
openai==1.12.0

# 限流中间件
slowapi==0.1.9

# 环境变量
python-dotenv==1.0.1